// src/app/business-logic/model/models/modelsUpdateRequest.ts
/**
 * models
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.7
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ModelsUpdateRequest {
    /**
     * Model id
     */
    model: string;
    /**
     * Model name Unique within the company.
     */
    name?: string;
    /**
     * Model comment
     */
    comment?: string;
    /**
     * User-defined tags
     */
    tags?: Array<string>;
    /**
     * System tags. This field is reserved for system use, please don\'t use it.
     */
    system_tags?: Array<string>;
    /**
     * Indication if the model is final and can be used by other tasks Default is   false.
     */
    ready?: boolean;
    /**
     * Model creation time (UTC)
     */
    created?: Date;
    /**
     * UI cache for this model
     */
    ui_cache?: object;
    /**
     * Project to which to model belongs
     */
    project?: string;
    /**
     * Associated task ID
     */
    task?: string;
    /**
     * Iteration (used to update task statistics if an associated task is reported)
     */
    iteration?: number;
}